more untangling of concerns
