+++ src/dom/ViewNode.ts
@@ -1,8 +1,4 @@
 
-import * as viewUtil from './utils'
-import * as types from 'tns-core-modules/utils/types'
-import { isAndroid, isIOS } from 'tns-core-modules/platform'
-import { View, EventData } from 'tns-core-modules/ui/core/view/view';
 import DocumentNode from './DocumentNode';
 
 
@@ -13,28 +9,12 @@ export function normalizeElementName(elementName: string) {
     .toLowerCase()}`
 }
 
-export interface ComponentMeta {
-  skipAddToDom?: boolean
-  insertChild?: (parent: ViewNode, child: ViewNode, index: number) => void;
-  removeChild?: (parent: ViewNode, child: ViewNode) => void;
-}
-
-export type EventListener = (args: any) => void;
-
 export function* elementIterator(el:ViewNode):Iterable<ViewNode> {
   yield el;
   for (let child of el.childNodes) {
      yield* elementIterator(child)
   }
 }
-
-
-const XML_ATTRIBUTES = Object.freeze([
-  'style',
-  'rows',
-  'columns',
-  'fontAttributes'
-])
 
 export default class ViewNode {
   nodeType: number;
@@ -44,8 +24,7 @@ export default class ViewNode {
   prevSibling: ViewNode;
   nextSibling: ViewNode;
   _ownerDocument: DocumentNode;
-  _nativeView: View;
-  _meta: ComponentMeta;
+  _attributes: {[index: string]: any };
 
   constructor() {
     this.nodeType = null
@@ -56,16 +35,15 @@ export default class ViewNode {
     this.nextSibling = null
 
     this._ownerDocument = null
-    this._nativeView = null
-    this._meta = null
+    this._attributes = {};  
   }
 
-  hasAttribute() {
-    return false
+  hasAttribute(name: string) {
+    return Object.keys(this._attributes).indexOf(name) > -1;
   }
 
-  removeAttribute() {
-    return false;
+  removeAttribute(name: string) {
+    delete this._attributes[name];
   }
 
   /* istanbul ignore next */
@@ -91,22 +69,6 @@ export default class ViewNode {
       : null
   }
 
-  get nativeView() {
-    return this._nativeView
-  }
-
-  set nativeView(view) {
-    if (this._nativeView) {
-      throw new Error(`Can't override native view.`)
-    }
-
-    this._nativeView = view
-  }
-
-  get meta() {
-     return this._meta
-  }
-
   /* istanbul ignore next */
   get ownerDocument():DocumentNode {
     if (this._ownerDocument) {
@@ -121,63 +83,13 @@ export default class ViewNode {
     return (this._ownerDocument = el as DocumentNode)
   }
 
-  getAttribute(key:string) {
-    return (this.nativeView as any)[key]
+  getAttribute(key:string):any {
+     return null;
   }
 
   /* istanbul ignore next */
   setAttribute(key:string, value:any) {
-    const nv = this.nativeView as any
-
-    if (!nv) return;
-
-    // normalize key
-    if (isAndroid && key.startsWith('android:')) {
-       key = key.substr(8);
-    }
-    if (isIOS && key.startsWith('ios:')) {
-       key = key.substr(4);
-    }
-    // try to fix case
-    let lowerkey = key.toLowerCase();
-    for (let realKey in nv) {
-       if (lowerkey == realKey.toLowerCase()) {
-         key = realKey;
-         break;
-       }
-    }
-    console.log(`setAttr ${this} ${key} ${value}`)
-    try {
-      if (XML_ATTRIBUTES.indexOf(key) !== -1) {
-        nv[key] = value
-      } else {
-        // detect expandable attrs for boolean values
-        // See https://vuejs.org/v2/guide/components-props.html#Passing-a-Boolean
-        if (types.isBoolean(nv[key]) && value === '') {
-          value = true
-        }
-        else {
-          nv[key] = value
-        }
-      }
-    } catch (e) {
-      // ignore but log
-      console.warn(`set attribute threw an error, attr:${key} on ${this._tagName}: ${e.message}`)
-    }
-  }
-
-  /* istanbul ignore next */
-  setStyle(property:string, value:string) {
-    console.log(`setStyle ${this} ${property} ${value}`)
-    if (!(value = value.trim()).length) {
-      return
-    }
-
-    if (property.endsWith('Align')) {
-      // NativeScript uses Alignment instead of Align, this ensures that text-align works
-      property += 'ment'
-    }
-    (this.nativeView.style as any)[property] = value
+     console.error("setting attribute on non native node")
   }
 
   /* istanbul ignore next */
@@ -190,17 +102,9 @@ export default class ViewNode {
     }
   }
 
-  /* istanbul ignore next */
-  addEventListener(event:string, handler: EventListener) {
-    console.log(`add event listener ${this} ${event}`)
-    this.nativeView.on(event, handler)
-  }
+  onInsertedChild( childNode: ViewNode, index: number) { }
 
-  /* istanbul ignore next */
-  removeEventListener(event:string, handler?: EventListener) {
-    console.log(`remove event listener ${this} ${event}`)
-    this.nativeView.off(event, handler)
-  }
+  onRemovedChild( childNode: ViewNode ) { }
 
   insertBefore(childNode:ViewNode, referenceNode:ViewNode) {
     console.log(`insert before ${this} ${childNode} ${referenceNode}`)
@@ -243,7 +147,7 @@ export default class ViewNode {
     referenceNode.prevSibling = childNode
     this.childNodes.splice(index, 0, childNode)
 
-    viewUtil.insertChild(this, childNode, index)
+    this.onInsertedChild(childNode, index);
   }
 
   appendChild(childNode:ViewNode) {
@@ -274,8 +178,7 @@ export default class ViewNode {
     }
 
     this.childNodes.push(childNode)
-
-    viewUtil.insertChild(this, childNode, this.childNodes.length - 1)
+    this.onInsertedChild(childNode, this.childNodes.length - 1)
   }
 
   removeChild(childNode:ViewNode) {
@@ -309,8 +212,7 @@ export default class ViewNode {
     childNode.nextSibling = null
 
     this.childNodes = this.childNodes.filter(node => node !== childNode)
-
-    viewUtil.removeChild(this, childNode)
+    this.onRemovedChild(childNode);
   }
 
   firstElement() {
@@ -320,13 +222,5 @@ export default class ViewNode {
       }
     }
     return null;
-  }
-
-  dispatchEvent(event: EventData) {
-    if (this.nativeView) { 
-       //nativescript uses the EventName while dom uses Type
-       event.eventName = (event as any).type;
-       this.nativeView.notify(event);
-    }
   }
 }
