+++ src/dom/utils.ts
@@ -2,6 +2,7 @@ import { LayoutBase } from 'tns-core-modules/ui/layouts/layout-base'
 import { ContentView } from 'tns-core-modules/ui/content-view'
 import { View }  from 'tns-core-modules/ui/core/view'
 import ViewNode from './ViewNode';
+import NativeElementNode from './NativeElementNode';
 
 
 export function isView(view: any) {
@@ -21,36 +22,17 @@ export function insertChild(parentNode: ViewNode, childNode:ViewNode, atIndex = 
     return
   }
 
+  if (!(parentNode instanceof NativeElementNode) || !(childNode instanceof NativeElementNode)) {
+     return
+  } 
+
   if (parentNode.meta && typeof parentNode.meta.insertChild === 'function') {
     return parentNode.meta.insertChild(parentNode, childNode, atIndex)
   }
 
-  if (!parentNode.nativeView || !childNode.nativeView) {
-     return
-  } 
-
   const parentView = parentNode.nativeView
   const childView = childNode.nativeView
 
-  /*if (parentView instanceof LayoutBase) {
-    if (childView.parent === parentView) {
-      let index = parentView.getChildIndex(childView)
-      if (index !== -1) {
-        parentView.removeChild(childView)
-      }
-    }
-    if (atIndex !== -1) {
-      parentView.insertChild(childView, atIndex)
-    } else {
-      parentView.addChild(childView)
-    }
-  } else if (parentView instanceof ContentView) {
-    if (childNode.nodeType === 8) {
-      parentView._addView(childView, atIndex)
-    } else {
-      parentView.content = childView
-    }
-  } else */
   if (parentView && (parentView as any)._addChildFromBuilder) {
     (parentView as any)._addChildFromBuilder(
       childNode._nativeView.constructor.name,
@@ -65,6 +47,10 @@ export function removeChild(parentNode: ViewNode, childNode: ViewNode) {
   if (!parentNode) {
     return
   }
+  
+  if (!(parentNode instanceof NativeElementNode) || !(childNode instanceof NativeElementNode)) {
+    return
+ } 
 
   if (parentNode.meta && typeof parentNode.meta.removeChild === 'function') {
     return parentNode.meta.removeChild(parentNode, childNode)
